Player camera improvements
Within this small update I applied a simple script that is applied to the camera, this will track the player. This script can allow me as a developer to change the offset which means i can have the camera at an distance away or on any angle. I also have smoothed the camera speed to counter any jitteriness it could encounter without it.