Music Persistence
In this version only a few minor additions were made. I selected and added a backing track used throughout all 3 levels. This track will also remain persistent and not restart at the beginning of each new level this is done through a script attached to the audio source. Finally there is a small white square acting as a means of testing if the music is remaining persistent through levels, after testing I have found this is the case.