Enemy Attacking
Within this next update I added the functionality for the enemy to attack the player. The enemy will project a ray cast out in front of them which will detect for the player, when it does so it will attack. The enemy has a cool down between attacks similarly to how I did so for the player, once the cool down has ended they can attack again.