Menus
Within this update I focussed on developing and linking menus within the game. This includes: Main Menu and defeat screen for each of the levels, victory screens are to come later. The main menu welcomes the player with a theme in the background, title of the game(Working progress) along with my University logo and of course the option to start the game or exit. As for the defeat screens they play a small game over jingle and the player is offered to either exit the game or restart the level they were just playing. So, if for example the player is beat on the second level and they want to restart then they will restart from the second level.