Boss Functionality
Within this update I created and developed the boss. It has a radius and if the player steps in front of the boss it will approach them while attacking. Once it has reached a certain distance away it will stop and continue attacking. To counter this the player can jump over and attack from behind after 5 seconds have passed and boss will turn facing the player again and resume attacking.